# coupon-api-server/src/main/java/com/github/dhslrl321/coupon/model/AssignCouponRequest.java
package com.github.dhslrl321.coupon.model;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class AssignCouponRequest {
    String code;
}