# coupon-api-server/src/main/java/com/github/dhslrl321/coupon/model/UseCouponSuccessResponse.java
package com.github.dhslrl321.coupon.model;

import java.time.LocalDateTime;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class UseCouponSuccessResponse {
    Long couponId;
    String code;
    LocalDateTime usedAt;
}